Разбил все неразбитые модули: в каждом из них по одному методу, все работает как надо
