+++ src/view/trip-days.js
@@ -2,16 +2,4 @@ const createTripDaysTemplate = () => {
   return `<ul class="trip-days"></ul>`;
 };
 
-const createDayTemplate = () => {
-  return `
-    <li class="trip-days__item  day">
-      <div class="day__info">
-        <span class="day__counter">1</span>
-        <time class="day__date" datetime="2019-03-18">MAR 18</time>
-      </div>
-
-      <ul class="trip-events__list"></ul>
-    </li>`;
-};
-
-export {createTripDaysTemplate, createDayTemplate};
+export {createTripDaysTemplate};
