+++ src/view/trip-event-details.js
@@ -1,0 +1,7 @@
+const createEventDetailsTemplate = () => {
+  return (
+    `<section class="event__details"></section>`
+  );
+};
+
+export {createEventDetailsTemplate};
