+++ src/view/trip-info-cost.js
@@ -1,0 +1,9 @@
+const createTripInfoCostTemplate = () => {
+  return (
+    `<p class="trip-info__cost">
+      Total: €&nbsp;<span class="trip-info__cost-value">1230</span>
+    </p>`
+  );
+};
+
+export {createTripInfoCostTemplate};
