+++ src/view/trip-info-main.js
@@ -1,0 +1,11 @@
+const createTripInfoMainTemplate = () => {
+  return (
+    `<div class="trip-info__main">
+      <h1 class="trip-info__title">Amsterdam — Chamonix — Geneva</h1>
+
+      <p class="trip-info__dates">Mar 18&nbsp;—&nbsp;20</p>
+    </div>`
+  );
+};
+
+export {createTripInfoMainTemplate};
